Move most of of bin into lib/cli.js, split out arg parsing.
